refactor: split nvmctrl into avr128 and tinyavr submodules
Replaces inline section-header comments with proper module structure,
matching the existing clock module layout.
